+++ packages/fiori-mcp-server/src/tools/run-rta-workflow-step/browser/index.ts
@@ -1,2 +1,2 @@
 export { callFrontendAction, defaultTransport, disconnectSite, stopBrowser } from './playwright-bridge';
-export type { FrontendActionResult, FrontendActionTransport } from './playwright-bridge';
+export type { FrontendActionResult, FrontendActionTransport } from './types';
